feat: Grid View on landing
With this we introduce a new grid view on the landing page as soon as
the viewport exceeds a certain point.

In the mobile view we do not show the potential post descriptions,
   although on the grid view (within media query) we do.

The grid layout is build out of four columns wher the last two posts in
the list have a column-span of two, the remaining items take one cell
each.